OPT.API: Major refactor of optics module.
Transform Optics module into a package.
Break down contents in modules:
 - acceptances
 - beam_envelope
 - edwards_teng
 - miscellaneous
 - rad_integrals
 - twiss

Delete methods:
 - calc_emittance_coupling
 - get_beam_size
 - get_traces

Rename methods:
 - EquilibriumParametersIntegrals --> EqParamsFromRadIntegrals
 - EquilibriumParametersOhmiFormalism --> EqParamsFromBeamEnvelope
 - calc_ohmienvelope --> calc_beamenvelope
